Continue refactoring the examples
Mostly putting stuff in separate files
and rearranging functions
