+++ examples/notebook/main.rs
@@ -1,46 +1,19 @@
-use gtk::glib;
+mod notebook;
+
 use gtk::prelude::*;
-use gtk::{IconSize, Orientation, ReliefStyle, Widget};
 
-struct Notebook {
-    notebook: gtk::Notebook,
-    tabs: Vec<gtk::Box>,
-}
+use notebook::Notebook;
 
-impl Notebook {
-    fn new() -> Notebook {
-        Notebook {
-            notebook: gtk::Notebook::new(),
-            tabs: Vec::new(),
-        }
-    }
+fn main() {
+    let application = gtk::Application::new(
+        Some("com.github.gtk-rs.examples.notebook"),
+        Default::default(),
+    )
+    .expect("Initialization failed...");
 
-    fn create_tab(&mut self, title: &str, widget: Widget) -> u32 {
-        let close_image = gtk::Image::from_icon_name(Some("window-close"), IconSize::Button);
-        let button = gtk::Button::new();
-        let label = gtk::Label::new(Some(title));
-        let tab = gtk::Box::new(Orientation::Horizontal, 0);
+    application.connect_activate(build_ui);
 
-        button.set_relief(ReliefStyle::None);
-        button.add(&close_image);
-
-        tab.pack_start(&label, false, false, 0);
-        tab.pack_start(&button, false, false, 0);
-        tab.show_all();
-
-        let index = self.notebook.append_page(&widget, Some(&tab));
-
-        button.connect_clicked(glib::clone!(@weak self.notebook as notebook => move |_| {
-            let index = notebook
-                .page_num(&widget)
-                .expect("Couldn't get page_num from notebook");
-            notebook.remove_page(Some(index));
-        }));
-
-        self.tabs.push(tab);
-
-        index
-    }
+    application.run();
 }
 
 fn build_ui(application: &gtk::Application) {
@@ -60,16 +33,4 @@ fn build_ui(application: &gtk::Application) {
 
     window.add(&notebook.notebook);
     window.show_all();
-}
-
-fn main() {
-    let application = gtk::Application::new(
-        Some("com.github.gtk-rs.examples.notebook"),
-        Default::default(),
-    )
-    .expect("Initialization failed...");
-
-    application.connect_activate(build_ui);
-
-    application.run();
 }
