+++ examples/overlay/style.css
@@ -1,0 +1,5 @@
+#overlay-label {
+    background-color: rgba(192, 192, 192, 0.8);
+    color: black;
+    font-weight: bold;
+}
